separate recursive bisection algorithm into own file
