+++ mt-kahypar/partition/CMakeLists.txt
@@ -10,6 +10,7 @@ set(PartitionSources
         context.cpp
         context_enum_classes.cpp
         metrics.cpp
+        recursive_bisection.cpp
         )
 
 foreach(modtarget IN LISTS TARGETS_WANTING_ALL_SOURCES)
